requirements.txt
dataframe_image==0.1.10
datasets==2.3.2
diffusers==0.12.1
matplotlib==3.7.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.5.0
requests==2.28.1
torch==1.12.0
transformers==4.27.1
